feat(infra): move provider selection to be owned by services
