+++ packages/application-generic/src/services/in-memory-provider/providers/providers.spec.ts
@@ -1,8 +1,6 @@
 import { IElasticacheClusterProviderConfig } from './elasticache-cluster-provider';
-import { IRedisClusterProviderConfig } from './redis-cluster-provider';
 import { IMemoryDbClusterProviderConfig } from './memory-db-cluster-provider';
-
-import { getClientAndConfigForCluster } from './index';
+import { getClusterProvider } from './providers';
 
 import { InMemoryProviderEnum } from '../types';
 
@@ -22,9 +20,7 @@ describe('Client and config for cluster', () => {
     process.env.REDIS_CLUSTER_SERVICE_HOST = redisClusterUrl;
     process.env.REDIS_CLUSTER_SERVICE_PORTS = redisClusterPorts;
 
-    const { getConfig } = getClientAndConfigForCluster(
-      InMemoryProviderEnum.ELASTICACHE
-    );
+    const { getConfig } = getClusterProvider(InMemoryProviderEnum.ELASTICACHE);
     const config: IElasticacheClusterProviderConfig = getConfig();
     expect(config.host).toEqual(elasticacheUrl);
     expect(config.port).toEqual(Number(elasticachePort));
@@ -37,84 +33,28 @@ describe('Client and config for cluster', () => {
     process.env.REDIS_CLUSTER_SERVICE_HOST = redisClusterUrl;
     process.env.REDIS_CLUSTER_SERVICE_PORTS = redisClusterPorts;
 
-    const { getConfig } = getClientAndConfigForCluster(
-      InMemoryProviderEnum.MEMORY_DB
-    );
+    const { getConfig } = getClusterProvider(InMemoryProviderEnum.MEMORY_DB);
     const config: IMemoryDbClusterProviderConfig = getConfig();
     expect(config.host).toEqual(memoryDbUrl);
     expect(config.port).toEqual(Number(memoryDbPort));
     expect(config.ttl).toEqual(7200);
   });
 
-  it('should return Redis Cluster config after validating Elasticache faulty URL config', () => {
-    process.env.ELASTICACHE_CLUSTER_SERVICE_HOST = '';
-    process.env.ELASTICACHE_CLUSTER_SERVICE_PORT = elasticachePort;
-    process.env.REDIS_CLUSTER_SERVICE_HOST = redisClusterUrl;
-    process.env.REDIS_CLUSTER_SERVICE_PORTS = redisClusterPorts;
-
-    const { getConfig } = getClientAndConfigForCluster(
-      InMemoryProviderEnum.ELASTICACHE
-    );
-    const config: IRedisClusterProviderConfig = getConfig();
-    expect(config.host).toEqual(redisClusterUrl);
-    expect(config.ports).toEqual(JSON.parse(redisClusterPorts));
-    expect(config.ttl).toEqual(7200);
-  });
-
-  it('should return Redis Cluster config after validating Elasticache faulty port config', () => {
-    process.env.ELASTICACHE_CLUSTER_SERVICE_HOST = elasticacheUrl;
-    process.env.ELASTICACHE_CLUSTER_SERVICE_PORT = '';
-    process.env.REDIS_CLUSTER_SERVICE_HOST = redisClusterUrl;
-    process.env.REDIS_CLUSTER_SERVICE_PORTS = redisClusterPorts;
-
-    const { getConfig } = getClientAndConfigForCluster(
-      InMemoryProviderEnum.ELASTICACHE
-    );
-
-    const config: IRedisClusterProviderConfig = getConfig();
-    expect(config.host).toEqual(redisClusterUrl);
-    expect(config.ports).toEqual(JSON.parse(redisClusterPorts));
-    expect(config.ttl).toEqual(7200);
-  });
-
-  it('should throw an error if Redis Cluster config has faulty URL config', () => {
+  it('should throw an error if Redis is passed as provider', () => {
     process.env.ELASTICACHE_CLUSTER_SERVICE_HOST = '';
     process.env.ELASTICACHE_CLUSTER_SERVICE_PORT = '';
     process.env.REDIS_CLUSTER_SERVICE_HOST = '';
     process.env.REDIS_CLUSTER_SERVICE_PORTS = redisClusterPorts;
 
     try {
-      const { getConfig } = getClientAndConfigForCluster(
-        InMemoryProviderEnum.ELASTICACHE
-      );
+      const { getConfig } = getClusterProvider(InMemoryProviderEnum.REDIS);
 
       fail('should not reach here');
     } catch (error) {
       expect(error).toBeInstanceOf(Error);
       const { message } = error as Error;
       expect(message).toEqual(
-        'Provider Elasticache is not properly configured in the environment variables'
-      );
-    }
-  });
-
-  it('should throw an error if Redis Cluster config has faulty port config', () => {
-    process.env.ELASTICACHE_CLUSTER_SERVICE_HOST = '';
-    process.env.ELASTICACHE_CLUSTER_SERVICE_PORT = '';
-    process.env.REDIS_CLUSTER_SERVICE_HOST = redisClusterUrl;
-    process.env.REDIS_CLUSTER_SERVICE_PORTS = '';
-
-    try {
-      const { getConfig } = getClientAndConfigForCluster(
-        InMemoryProviderEnum.ELASTICACHE
-      );
-
-      fail('should not reach here');
-    } catch (error) {
-      expect(error).toBeInstanceOf(Error);
-      const { message } = error as Error;
-      expect(message).toEqual(
-        'Provider Elasticache is not properly configured in the environment variables'
+        'Provider Redis is not available in Cluster Mode'
       );
     }
   });
